Add LauncherUtil class, menu item for alter launcher
