+++ app/src/main/res/values/strings.xml
@@ -32,7 +32,8 @@
 
     <!-- Menu Strings -->
     <string name="menu_title_about">About</string>
-    <string name="menu_title_reset_defaults">Reset Defaults</string>
+    <string name="menu_title_reset_defaults">Reset Default Settings</string>
+    <string name="menu_title_alter_default_launcher">Alter Default Launcher</string>
     <string name="menu_title_show_apps">Show Apps</string>
     <string name="menu_app_open">Open</string>
     <string name="menu_app_uninstall">Uninstall</string>
